+++ dataFormatter/format_test.go
@@ -1,13 +1,14 @@
-package main
+package dataformatter
 
 import (
+	"go-qiita/types"
 	"testing"
 )
 
 func TestConvertToArticleCsv(t *testing.T) {
-    var data []Article
-    test1 := Article{Title: "test 1", Url: "https://test.com/1"}
-    test2 := Article{Title: "test 2", Url: "https://test.com/2"}
+    var data []types.Article
+    test1 := types.Article{Title: "test 1", Url: "https://test.com/1"}
+    test2 := types.Article{Title: "test 2", Url: "https://test.com/2"}
     data = append(data, test1, test2)
     csvData := ConvertToCsv(data)
     header := csvData[0]
@@ -19,7 +20,4 @@ func TestConvertToArticleCsv(t *testing.T) {
     if len(csvData) == 1 {
         t.Error("変換データがありません")
     }
-    
-
 }
-
