+++ types/article.go
@@ -1,0 +1,6 @@
+package types
+
+type Article struct {
+	Title string `json:"title"`
+	Url   string `json:"url"`
+}
